1.1
Changed how casting is calculated. Previously, the calculation was based of points, now it utilizes vector cross products to calculate intersection points. Added a dictionary system to keep track of any and walls for the rays to collide with. Added a system to check if a wall has been collided with to reduce the amount of ray calculations needed.